requirements.txt
asgiref==3.8.1; python_version >= '3.8'
django==5.0.5; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytz==2024.1
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.11.0; python_version < '3.11'
tzdata==2024.1; python_version >= '2'
